+++ src/Catty/ViewController/Download/Details/ProjectDetailStoreViewControllerNewDownloadExtension.swift
@@ -24,64 +24,5 @@ import Foundation
 
 
 extension ProjectDetailStoreViewController {
-    func download(name: String) {
-        storeProjectDownloader.download(
-            projectId: self.project.projectID,
-            projectName: name,
-            completion: { _, storeProjectDownloaderError in
-                if let error = storeProjectDownloaderError {
-                    switch error {
-                    case .cancelled:
-                        return
-                    case .unexpectedError, .timeout:
-                        Util.defaultAlertForNetworkError()
-                    case .parse(error: _), .request(error: _, statusCode: _):
-                        Util.alert(text: kLocalizedInvalidZip)
-                    }
 
-                    self.resetDownloadStatus()
-                    return
-                }
-
-                self.downloadFinished()
-
-            }, progression: { progress in
-                self.updateProgress(Double(progress))
-                self.reloadInputViews()
-            })
-    }
-
-    private func downloadFinished() {
-        self.project.isdownloading = false
-
-        if let button = self.view.viewWithTag(Int(kStopLoadingTag)) as? EVCircularProgressView {
-            button.isHidden = true
-            button.progress = 0
-        }
-
-        if let openButton = self.view.viewWithTag(Int(kOpenButtonTag)) {
-            openButton.isHidden = false
-        }
-
-        if let downloadButton = self.view.viewWithTag(Int(kDownloadAgainButtonTag)) as? UIButton {
-            downloadButton.isEnabled = true
-            downloadButton.isHidden = false
-        }
-
-        Util.setNetworkActivityIndicator(false)
-    }
-
-    private func resetDownloadStatus() {
-        self.view.viewWithTag(Int(kDownloadButtonTag))?.isHidden = false
-        self.view.viewWithTag(Int(kOpenButtonTag))?.isHidden = true
-        self.view.viewWithTag(Int(kStopLoadingTag))?.isHidden = true
-        self.view.viewWithTag(Int(kDownloadAgainButtonTag))?.isHidden = true
-
-        Util.setNetworkActivityIndicator(false)
-    }
-
-    private func updateProgress(_ progress: Double) {
-        guard let button = self.view.viewWithTag(Int(kStopLoadingTag)) as? EVCircularProgressView else { return }
-        button.setProgress(progress, animated: true)
-    }
 }
